requirements.txt
# Requisitos para scraping de Binance P2P y aplicación web
# Autor: GitHub Copilot
# Fecha: Mayo 2025

# HTTP requests
requests>=2.25.1

# Base de datos PostgreSQL
psycopg2-binary>=2.9.0

# Aplicación web
Flask>=2.0.1
